Run the codeflow code check as it's own job
