+++ src/Microsoft.DotNet.Darc/DarcLib/VirtualMonoRepo/VmrPatchHandler.cs
@@ -348,7 +348,7 @@ public class VmrPatchHandler : IVmrPatchHandler
             }
         }
 
-        _logger.LogDebug("{output}", result.ToString());
+        //_logger.LogDebug("{output}", result.ToString());
 
         if (removePatchAfter)
         {
